+++ sdk/cpp/CHANGES.md
@@ -7,7 +7,7 @@
 #### Resolved GitHub issues
   * Fix jsonSubtreeCodec only serializing up to 4B numbers
   * Fixed bug in extracting module name from namespace
-
+  * gNMI set RPC fails when JSON payload is formatted string
 
 ### 2019-10-15 version 0.8.4